+++ Assets/Dust/CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Added
 
 - Added component Animation.DuScale
+- Added DuActions [core]
+- Added DuActionMoveBy & DuActionMoveTo
 
 ### Fixed